Move out the custom gen feature login
